+++ config/defines.php
@@ -1,6 +1,6 @@
 <?php
 
-define('ANTRAGSGRUEN_VERSION', '4.0.3');
+define('ANTRAGSGRUEN_VERSION', '4.0.4');
 define('ANTRAGSGRUEN_HISTORY_URL', 'https://github.com/CatoTH/antragsgruen/blob/master/History.md');
 define('ANTRAGSGRUEN_UPDATE_BASE', 'https://antragsgruen.de/updates/');
 define('ANTRAGSGRUEN_MIN_PHP_VERSION', '5.6.0');