+++ History.md
@@ -4,6 +4,7 @@
 
 - Bugfix: Creating the proposed procedure list could fail.
 - Bugfix: A ODS-list could not be exported if a motion had no initiators.
+- Bugfix: Another bug regarding nested list items and line numbers in PDF generation was fixed.
 
 ### Version 4.0.3 (2018-10-14)